altered shop script to enable player
player was previously disabled when shop was open in order for player input to be avoided in the menu, but this would prevent items from directly being transferred to the inventory, which is part of the player prefab. i worked around this by basically treating the shop menu the same way as the inventory (stopping time) and manually disabling the ability to shoot through a bool.